+++ src/main/java/com/pro/infomate/approval/dto/request/DocumentRequest.java
@@ -22,7 +22,7 @@ public class DocumentRequest {
   private String content; // 문서 내용
   private String emergency; // 긴급여부
 
-  private List<Long> refList; // 참조자 멤버목록(코드만)
+  private List<Integer> refList; // 참조자 멤버목록(코드만)
 
   private List<ApprovalRequest> approvalList;// 결재 목록
