+++ src/main/java/com/pro/infomate/member/repository/MemberRepository.java
@@ -7,12 +7,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
 
   @EntityGraph(attributePaths = {"department", "rank"})
   Member findByMemberId(long memberId);
 
-  List<Member> findByMemberCodeIn(List<Long> ids);
+  List<Member> findByMemberCodeIn(List<Integer> ids);
 
-  Member findByMemberCode(Long memberCode);
+  Member findByMemberCode(int memberCode);
 }
